+++ vpnmanager/impl/src/main/java/org/opendaylight/netvirt/vpnmanager/iplearn/LearntVpnVipToPortEventProcessor.java
@@ -52,6 +52,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.l3vpn.rev130911.adj
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.l3vpn.rev130911.learnt.vpn.vip.to.port.event.data.LearntVpnVipToPortEvent;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.neutronvpn.rev150602.neutron.vpn.portip.port.data.VpnPortipToPort;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint32;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -227,9 +228,9 @@ public class LearntVpnVipToPortEventProcessor
                     }
 
                     String rd = vpnUtil.getVpnRd(vpnInstName);
-                    long label = vpnUtil.getUniqueId(VpnConstants.VPN_IDPOOL_NAME,
+                    Uint32 label = vpnUtil.getUniqueId(VpnConstants.VPN_IDPOOL_NAME,
                         VpnUtil.getNextHopLabelKey(rd != null ? rd : vpnInstName, ip));
-                    if (label == 0) {
+                    if (label.longValue() == 0) {
                         LOG.error("Unable to fetch label from Id Manager. Bailing out of adding MIP adjacency {}"
                                 + " to vpn interface {} for vpn {}", ip, vpnInterface, vpnInstName);
                         return;
